FFM-5306 - Update ff-test-cases for prereq id/value mismatch
What
Update test cases to include a test for prereq using different ids and values.
string-flag-with-dependency feature has a dependency on feature-flag-with-mixed-id-and-values which will trigger the bug.

Why
We're incorrectly matching values when we should be using identifiers. This can cause features not to trigger.

Testing
Tested with Java SDK integration tests